Feature/etapa 02

+++ services/reservas-svc/src/main/java/edu/utn/frsf/isi/dan/reservas_svc/model/Habitacion.java
@@ -15,7 +15,7 @@ import java.util.List;
 public class Habitacion {
     @Id
     private String id;
-    private Long habitacionId;
+    private Integer habitacionId;
     private Integer capacidad;
     private Double precioNoche;
     private List<String> amenities;
