+++ services/user-svc/src/main/java/edu/utn/frsf/isi/dan/user/dto/PropietarioDTOResponse.java
@@ -7,7 +7,7 @@ public record PropietarioDTOResponse(
         String telefono,
         String dni,
         CuentaBancariaDTOResponse cuentaBancaria,
-        Long idHotel
+        Integer idHotel
 ) {
 
 }
